+++ lib/widgets/misc/tasks_counts.dart
@@ -69,12 +69,7 @@ class TasksCountWidget extends StatelessWidget {
             padding: const EdgeInsets.symmetric(horizontal: 4),
             child: Text(
               '$label: ${snapshot.data}',
-              style: TextStyle(
-                color: styleConfig().primaryTextColor,
-                fontFamily: 'Oswald',
-                fontSize: 12,
-                fontWeight: FontWeight.w100,
-              ),
+              style: searchLabelStyle(),
             ),
           );
         }