+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Decluttered task view: only show editor toolbar on first focus
 - Decluttered task view: only show tags in task comments when not equal to parent tags
+- Show total time spent on a task
 
 ## [0.8.232] - 2023-01-13
 ### Changed: